backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
pypdf
python-multipart
python-dotenv
